Check for gpu hardware continually.
Instead of checking for gpus only once, check on every hook
invocation, or until cuda support is detected.

This change allows the layer to install cuda packages when gpu
hardware is added to an already-running unit. Before, since the
check was run only_once, the gpus had to be present at the time
the install hook was executed.