blacklist nouveau instead of removing packages
The opensource nouveau module conflicts with the nvidia bits that
get installed by this layer. Removing nouveau packages has the
unwanted side effect of removing lots of x11/java deps. This is
bad news for Bigtop charms as openjdk *must* remain on the system
to be of any value.

Instead of the apt remove/autoremove bits, blacklist the nouveau
module. Do this in a prereq function so the 'install_cuda' func
isn't so complex.